ghost car
i have made a change in my scene with the race track to make it more testable. i have also made the beginning of the ghost cars and the working of that part. it isnt finished yet but i have come a far way with it. still a lot of buggs tho